+++ assets/Add.6a053471.js
@@ -1,0 +1,1 @@
+import{t}from"./vendor.37d14b95.js";import{c as e}from"./index.5b3b4b7a.js";const o=t('<svg><path d="M0 0h24v24H0z" fill="none"></path></svg>',4,!0),v=t('<svg><path d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6v2z"></path></svg>',4,!0);var h=e(()=>[o.cloneNode(!0),v.cloneNode(!0)],"Add");export{h as A};