+++ assets/SkeletonPage.6478be16.js
@@ -1,0 +1,13 @@
+var o=Object.defineProperty;var a=(r,i)=>o(r,"name",{value:i,configurable:!0});import{a3 as n}from"./index.5b3b4b7a.js";import{C as m}from"./ComponentInfo.bede338e.js";import{S as c}from"./copyText.d14c4ac4.js";import{a as t}from"./vendor.37d14b95.js";import"./Link.9626be27.js";import"./PaperCode.cf41c2a7.js";function e(){return t(c,{spacing:1,get children(){return[t(n,{variant:"text"}),t(n,{variant:"circular",width:40,height:40}),t(n,{variant:"rectangular",width:210,height:118})]}})}a(e,"Variants");e.code=`import Skeleton from "@suid/material/Skeleton";
+import Stack from "@suid/material/Stack";
+
+export default function Variants() {
+  return (
+    <Stack spacing={1}>
+      <Skeleton variant="text" />
+      <Skeleton variant="circular" width={40} height={40} />
+      <Skeleton variant="rectangular" width={210} height={118} />
+    </Stack>
+  );
+}
+`;function s(){return t(m,{get name(){return n.name},examples:[e]})}a(s,"SkeletonPage");export{s as default};