+++ assets/useMediaQueryPage.ffdede16.js
@@ -1,0 +1,8 @@
+var m=Object.defineProperty;var n=(t,e)=>m(t,"name",{value:e,configurable:!0});import{e as a}from"./index.5b3b4b7a.js";import{C as i}from"./ComponentInfo.bede338e.js";import{r as o,t as s,a as p}from"./vendor.37d14b95.js";import"./Link.9626be27.js";import"./copyText.d14c4ac4.js";import"./PaperCode.cf41c2a7.js";const u=s("<span></span>");function r(){const t=a("(min-width:600px)");return(()=>{const e=u.cloneNode(!0);return o(e,()=>`(min-width:600px) matches: ${t()}`),e})()}n(r,"SimpleMediaQuery");r.code=`import useMediaQuery from "@suid/material/useMediaQuery";
+
+export default function SimpleMediaQuery() {
+  const matches = useMediaQuery("(min-width:600px)");
+
+  return <span>{\`(min-width:600px) matches: \${matches()}\`}</span>;
+}
+`;function M(){return p(i,{get name(){return a.name},docsApiName:"components/use-media-query",examples:[r],moreExamples:!1})}n(M,"useMediaQueryPage");export{M as default};