+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @suid/site
 
+## 0.0.24
+
+### Patch Changes
+
+- [`08c3233`](https://github.com/swordev/suid/commit/08c32330e3098864bc39f19745558ca455291db3) Thanks [@juanrgm](https://github.com/juanrgm)! - Add `Checkbox` component page
+
+- Updated dependencies [[`67076af`](https://github.com/swordev/suid/commit/67076af65e0bb7a8c9c4186a64ddf59c6262af78), [`b819235`](https://github.com/swordev/suid/commit/b819235dd3f4988d62492d0d4290043b3e00cc18), [`16dcdd3`](https://github.com/swordev/suid/commit/16dcdd341bb7c7148126046a848b81ed85ae210c)]:
+  - @suid/material@0.0.18
+  - @suid/system@0.0.15
+  - @suid/icons-material@0.0.18
+  - @suid/base@0.0.16
+
 ## 0.0.23
 
 ### Patch Changes